requirements.txt
scikit-image == 0.16.2      # Version Matching. Must be version 0.6.1
keras == 2.1.6              # Minimum version 4.1.1
h5py == 2.10.0             # Version Exclusion. Anything except version 3.5
tensorflow-cpu == 1.15.0
numpy == 1.21.5
matplotlib == 3.5.1
Django == 3.0.2
dj-static
Cython
pycocotools
django-cors-headers
Pillow == 9.0.0
gunicorn == 20.0.4
requests == 2.26.0
django_heroku

